containerd: minimize restarts
